Fix invalid translations
'msgid' and 'msgstr' entries do not both end with '\n' which is
mandatory according to specification from the 'msgfmt' manual:

```
The msgid and msgstr strings are studied and compared. It is considered
abnormal if one string starts or ends with a newline while the other
does not.
```

Besides all placeholders from the 'msgid' must be used in the 'mgstr'.